refactor(runtime): API consistency audit — normalize naming, expand runtime.def, add missing operations
Standardize collection API naming across all runtime types: put→add, drop→remove,
contains→has. Add missing operations (first/last/reverse/clone) to Ring, Queue, Stack,
Deque, and Set. Expand runtime.def with 221 new entries for full registry coverage.
Rename rt_ns_bridge → rt_sb_bridge. Update all tests, docs, and sweep files to match.

tests: all runtime, golden, and sweep tests updated for renamed APIs

+++ src/runtime/core/rt_context.c
@@ -167,7 +167,7 @@ void rt_set_main_thread(void)
     __atomic_store_n(&g_main_thread_set_, 1, __ATOMIC_RELEASE);
 }
 
-int rt_is_main_thread(void)
+int8_t rt_is_main_thread(void)
 {
     if (!__atomic_load_n(&g_main_thread_set_, __ATOMIC_ACQUIRE))
         return 1; // Before init, assume main thread (avoids false positives)
